64bit Server.exe PR

+++ server/NetworkServer.h
@@ -75,7 +75,7 @@ private:
 
 	int loopCount;
 
-	UINT offsets[OT_max];
+	QWORD offsets[OT_max];
 
 	bool quickInfo;
 
@@ -91,7 +91,7 @@ private:
 
 
 
-public:	
+public:
 
 	// Bit flags (in clientRequest) determining what data the client is requesting. See client's Structures.cs.
 
@@ -99,25 +99,25 @@ public:
 
 	{
 
-		IPT_zone =		0x01,
+		IPT_zone = 0x01,
 
-		IPT_self =		0x02,
+		IPT_self = 0x02,
 
-		IPT_target =	0x04,
+		IPT_target = 0x04,
 
-		IPT_spawns =	0x08,
+		IPT_spawns = 0x08,
 
-		IPT_ground =	0x10,
+		IPT_ground = 0x10,
 
-		IPT_getproc =	0x20,
+		IPT_getproc = 0x20,
 
-		IPT_setproc =	0x40,
+		IPT_setproc = 0x40,
 
-		IPT_world =		0x80
+		IPT_world = 0x80
 
 	};
 
-	
+
 
 	// The type of element we are sending back to client. Lives in the 'flags' field of the outgoing packets.
 
@@ -125,19 +125,19 @@ public:
 
 	{
 
-		OPT_spawns =	0x00,
+		OPT_spawns = 0x00,
 
-		OPT_target =	0x01,
+		OPT_target = 0x01,
 
-		OPT_zone =		0x04,
+		OPT_zone = 0x04,
 
-		OPT_ground =	0x05,
+		OPT_ground = 0x05,
 
-		OPT_process =	0x06,
+		OPT_process = 0x06,
 
-		OPT_world =		0x08,
+		OPT_world = 0x08,
 
-		OPT_self =		0xFD
+		OPT_self = 0xFD
 
 	};
 
@@ -160,9 +160,9 @@ public:
 
 	void closeClientSocket();
 
-	void setOffset(offset_types ot, UINT value);
+	void setOffset(offset_types ot, QWORD value);
 
-	void init(IniReaderInterface* ir_intf);	
+	void init(IniReaderInterface* ir_intf);
 
 	void enterReceiveLoop(MemReaderInterface* mr_intf);
 
@@ -175,4 +175,3 @@ public:
 	UINT current_offset(int type);
 
 };
-
